* dxbjavid/ayiya-identity-len-truncation:
  ayiya: add test for identity length that overflows uint8_t
  widen ayiya identity_len to avoid uint8_t truncation

(cherry picked from commit 5e98a79c31f79ce210044aad760b8b717386834e)

+++ src/packet_analysis/protocol/ayiya/AYIYA.cc
@@ -34,7 +34,7 @@ bool AYIYAAnalyzer::AnalyzePacket(size_t len, const uint8_t* data, Packet* packe
         return false;
     }
 
-    uint8_t identity_len = 1 << (data[0] >> 4);
+    size_t identity_len = 1 << (data[0] >> 4);
     uint8_t signature_len = (data[1] >> 4) * 4;
     hdr_size += identity_len + signature_len;
